Update SunriseAlarmClock.ino
Tested how the arduino millis() function works vs the clock on my computer, and built the hardware and attached to the device. Tested everything, and it seems to be working well. It would be nice to have a way to set the clock that is easier, but for now you can program the time and then reset the arduino at the correct time to make sure the time is synced.